# backend/cmd/docs/swagger.yaml
basePath: /
definitions:
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
externalDocs:
  description: OpenAPI Specification
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@faladev.com
    name: Suporte da API Faladev
    url: http://www.faladev.com/support
  description: Esta é uma API exemplo do Faladev, que integra com o Google Calendar
    e o Gmail.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Faladev API
  version: "1.0"
paths:
  /:
    get:
      description: Renders the form.html page to display user information form.
      produces:
      - text/html
      responses:
        "200":
          description: HTML content of the form page
          schema:
            type: string
      summary: Render form page
  /callback:
    get:
      consumes:
      - application/json
      description: Exchange code for token and save it
      parameters:
      - description: State token
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "303":
          description: Redirects to /
        "400":
          description: State token doesn't match
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Unable to retrieve or save token
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Handle OAuth2 callback
  /event:
    post:
      consumes:
      - application/json
      description: Handles event creation, guest addition, email sending, and redirects
        to Google Meet link.
      parameters:
      - description: Name of the student
        in: formData
        name: name
        required: true
        type: string
      - description: Email of the student
        in: formData
        name: email
        required: true
        type: string
      - description: Phone number of the student
        in: formData
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "303":
          description: Redirects to Google Meet link
          schema:
            type: string
        "400":
          description: No Google Meet link available or other errors
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Handle event creation and interaction
swagger: "2.0"